Move resource server to dedicated pkg
